Removed PolygonDialogInfoRepository.java and FeatureRepository updated with LocalValueStore.
