+++ gnd/src/main/java/com/google/android/gnd/MainViewModel.java
@@ -25,7 +25,6 @@ import androidx.navigation.NavDirections;
 import com.google.android.gnd.model.Project;
 import com.google.android.gnd.model.TermsOfService;
 import com.google.android.gnd.repository.FeatureRepository;
-import com.google.android.gnd.repository.PolygonDialogInfoRepository;
 import com.google.android.gnd.repository.ProjectRepository;
 import com.google.android.gnd.repository.TermsOfServiceRepository;
 import com.google.android.gnd.repository.UserRepository;
@@ -68,7 +67,6 @@ public class MainViewModel extends AbstractViewModel {
   private final FeatureRepository featureRepository;
   private final UserRepository userRepository;
   private final TermsOfServiceRepository termsOfServiceRepository;
-  private final PolygonDialogInfoRepository polygonDialogInfoRepository;
   private final EphemeralPopups popups;
 
   @Inject
@@ -77,7 +75,6 @@ public class MainViewModel extends AbstractViewModel {
       FeatureRepository featureRepository,
       UserRepository userRepository,
       TermsOfServiceRepository termsOfServiceRepository,
-      PolygonDialogInfoRepository polygonDialogInfoRepository,
       Navigator navigator,
       AuthenticationManager authenticationManager,
       EphemeralPopups popups,
@@ -85,7 +82,6 @@ public class MainViewModel extends AbstractViewModel {
     this.projectRepository = projectRepository;
     this.featureRepository = featureRepository;
     this.termsOfServiceRepository = termsOfServiceRepository;
-    this.polygonDialogInfoRepository = polygonDialogInfoRepository;
     this.userRepository = userRepository;
     this.popups = popups;
 
@@ -129,7 +125,7 @@ public class MainViewModel extends AbstractViewModel {
   private Observable<NavDirections> onSignInStateChange(SignInState signInState) {
     if (signInState.state() != State.SIGNED_IN) {
       termsOfServiceRepository.setTermsOfServiceAccepted(false);
-      polygonDialogInfoRepository.setPolygonDialogInfoShown(false);
+      featureRepository.setPolygonDialogInfoShown(false);
     }
 
     switch (signInState.state()) {
