+++ gnd/src/main/java/com/google/android/gnd/ui/home/mapcontainer/PolygonDrawingViewModel.java
@@ -30,7 +30,7 @@ import com.google.android.gnd.model.feature.Point;
 import com.google.android.gnd.model.feature.PolygonFeature;
 import com.google.android.gnd.model.layer.Layer;
 import com.google.android.gnd.persistence.uuid.OfflineUuidGenerator;
-import com.google.android.gnd.repository.PolygonDialogInfoRepository;
+import com.google.android.gnd.repository.FeatureRepository;
 import com.google.android.gnd.rx.BooleanOrError;
 import com.google.android.gnd.rx.Nil;
 import com.google.android.gnd.rx.annotations.Hot;
@@ -81,18 +81,18 @@ public class PolygonDrawingViewModel extends AbstractViewModel {
 
   private final OfflineUuidGenerator uuidGenerator;
   private final AuthenticationManager authManager;
-  private final PolygonDialogInfoRepository polygonDialogInfoRepository;
+  private final FeatureRepository featureRepository;
   @Nullable private Point cameraTarget;
 
   @Inject
   PolygonDrawingViewModel(
       LocationManager locationManager,
-      PolygonDialogInfoRepository polygonDialogInfoRepository,
+      FeatureRepository featureRepository,
       AuthenticationManager authManager,
       OfflineUuidGenerator uuidGenerator) {
     this.locationManager = locationManager;
     this.authManager = authManager;
-    this.polygonDialogInfoRepository = polygonDialogInfoRepository;
+    this.featureRepository = featureRepository;
     this.uuidGenerator = uuidGenerator;
     Flowable<BooleanOrError> locationLockStateFlowable = createLocationLockStateFlowable().share();
     this.locationLockState =
@@ -253,10 +253,10 @@ public class PolygonDrawingViewModel extends AbstractViewModel {
   }
 
   public boolean isPolygonInfoDialogShown() {
-    return polygonDialogInfoRepository.isPolygonDialogInfoShown();
+    return featureRepository.isPolygonDialogInfoShown();
   }
 
   public void updatePolygonInfoDialogShown() {
-    polygonDialogInfoRepository.setPolygonDialogInfoShown(true);
+    featureRepository.setPolygonDialogInfoShown(true);
   }
 }
